requirements.txt
argcomplete==2.0.0
argh==0.26.2
attrs==21.4.0
biopython==1.79
black==22.1.0
cachetools==5.0.0
click==8.0.4
gffutils==0.10.1
Jinja2==3.0.3
jsonlines==3.0.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
ncbi-datasets-pylib==13.8.0
numpy==1.22.2
pathspec==0.9.0
platformdirs==2.5.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyfaidx==0.6.4
python-dateutil==2.8.2
simplejson==3.17.6
six==1.16.0
tomli==2.0.1
urllib3==1.26.9
